+++ Happiggy-bank/Happiggy-bank/Utils/Constants.swift
@@ -104,119 +104,6 @@ extension ColorButton {
     }
 }
 
-extension HomeView {
-    
-    /// HomeView 에서 설정하는 layout에 적용할 상수값
-    enum Metric {
-        
-        /// 디바이스 스크린 높이
-        static let screenHeight: CGFloat = UIScreen.main.bounds.height
-        
-        /// 디바이스 스크린 너비
-        static let screenWidth: CGFloat = UIScreen.main.bounds.width
-    }
-}
-
-extension BottleListViewController {
-    
-    /// BottleListViewController에서 설정하는 layout 상수값
-    enum Metric {
-        
-        /// 컬렉션뷰 셀 너비
-        static let cellWidth: CGFloat = BottleCell.Metric.cellWidth
-        
-        /// 컬렉션뷰 셀 높이
-        static let cellHeight: CGFloat = BottleCell.Metric.cellHeight
-        
-        /// 셀 각 라인 별 최소 간격
-        static let minimumLineSpacing: CGFloat = 40
-    }
-    
-    /// BottleListViewController에서 설정하는 Color 상수값
-    enum Color {
-        
-        /// 내비게이션 바 Item 색상
-        static let navigationBar: Int = 0x000000
-    }
-    
-    /// BottleListViewController에서 설정하는 글자 크기 상수값
-    enum FontSize {
-        
-        /// 리스트가 빈 경우 표시되는 라벨 글자 크기
-        static let emptyListLabelTitle: CGFloat = 16
-    }
-    
-    /// BottleListViewController에서 설정하는 문자열
-    enum StringLiteral {
-        
-        /// 리스트가 빈 경우 표시되는 내비게이션 타이틀
-        static let emptyListNavigationBarTitle: String = "저금통 목록"
-        
-        /// 리스트가 차있는 경우 표시되는 내비게이션 타이틀
-        static let fullListNavigationBarTitle: String = "저금통 목록"
-        
-        /// 리스트가 빈 경우 테이블뷰에 표시되는 라벨
-        static let emptyListLabelTitle: String = "이전에 사용한 저금통이 없습니다."
-        
-        /// 쪽지 리스트로 넘어갔을 때 뒤로가기 버튼의 제목을 비워두기 위해 사용
-        static let emptyString = ""
-    }
-}
-
-extension BottleCell {
-    
-    /// 리유저블 ID 이름
-    static var reuseIdentifier: String {
-        return "\(self)"
-    }
-    
-    /// 인터페이스 빌더 파일 이름
-    static var nibName: String {
-        return "\(self)"
-    }
-    
-    /// Bottle Cell에서 설정하는 layout 상수값
-    enum Metric {
-        
-        /// bottle  image 가로:세로 비율
-        private static let imageSizeRatio: CGFloat = 306 / 165
-        
-        /// bottle cell 가로:세로 비율
-        private static let cellSizeRatio: CGFloat = 356 / 165
-        
-        /// 화면 가로 길이
-        static let screenWidth: CGFloat = UIScreen.main.bounds.width
-        
-        /// bottle cell 사이 수평 간격
-        static let cellHorizontalSpacing: CGFloat = 24
-        
-        /// bottle cell 사이 수직 간격
-        static let cellVerticalSpacing: CGFloat = 20
-        
-        /// bottle cell 너비
-        static let cellWidth: CGFloat = ((screenWidth - 3 * 24) / 2)
-        
-        /// bottle cell 높이
-        static let cellHeight = imageHeight + 50
-        
-        /// bottle image 높이 : 스크린 가로 길이에서 좌우 패딩값을 뺀 다음 지정한 비율을 곱해서 계산
-        static let imageHeight = cellWidth * imageSizeRatio
-        
-        /// corner radius
-        static let cornerRadius: CGFloat = 7
-    }
-    
-    /// Bottle Cell에서 설정하는 글자 크기
-    enum FontSize {
-        
-        /// 유리병 제목 라벨 글자 크기
-        static let titleLabel: CGFloat = 16
-        
-        /// 유리병 기간 라벨 글자 크기
-        static let dateLabel: CGFloat = 12
-    }
-}
-
 extension Bottle {
     
     /// Bottle 엔티티에서 설정하는 문자열
